# specs/001-user-authentication/contracts/auth.openapi.yaml
openapi: 3.0.3
info:
  title: Authentication API
  version: 1.0.0
  description: Contract for registration, login, password reset, and token validation behavior.
servers:
  - url: https://api.example.com
paths:
  /auth/register:
    post:
      summary: Register a new user account
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid request
        '409':
          description: Email already registered
  /auth/login:
    post:
      summary: Authenticate and issue JWT access token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
        '429':
          description: Rate limited
  /auth/password-reset/request:
    post:
      summary: Request password reset email
      operationId: requestPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestBody'
      responses:
        '202':
          description: Request accepted (generic non-enumerating response)
        '429':
          description: Rate limited
  /auth/password-reset/confirm:
    post:
      summary: Confirm password reset with one-time token
      operationId: confirmPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '200':
          description: Password reset completed and sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetConfirmResponse'
        '400':
          description: Invalid token or policy violation
        '410':
          description: Token expired or already used
        '429':
          description: Rate limited
  /auth/session/validate:
    get:
      summary: Validate current access token state
      operationId: validateSession
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token is active and valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionValidationResponse'
        '401':
          description: Token invalid, expired, or revoked
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          description: Must satisfy password policy (min 12 chars, upper/lower/number/symbol, not breached).
    RegisterResponse:
      type: object
      required: [userId, email]
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresInSeconds]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresInSeconds:
          type: integer
          enum: [86400]
    PasswordResetRequestBody:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    PasswordResetConfirmRequest:
      type: object
      required: [resetToken, newPassword]
      properties:
        resetToken:
          type: string
        newPassword:
          type: string
    PasswordResetConfirmResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [password_reset]
    SessionValidationResponse:
      type: object
      required: [active, expiresAt]
      properties:
        active:
          type: boolean
        expiresAt:
          type: string
          format: date-time